fixed a bug where the new score wont change in the drawer

+++ app/src/main/java/com/example/quadwrangle/game_view/ui/game/GameViewModel.java
@@ -165,7 +165,7 @@ public class GameViewModel extends AndroidViewModel {
     public boolean handleHighScore() {
         // high score only for AI mode.
         if (isAI.getValue()) { // if the game mode is AI
-            int scoreFromData = userDbLeaderboardConnector.getMyHighScore();
+            long scoreFromData = userDbLeaderboardConnector.getMyHighScore();
             int currentPlayer1Score = player1Score.getValue();
             if (currentPlayer1Score > scoreFromData) {
                 userDbLeaderboardConnector.updateScore(currentPlayer1Score);
